requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.14.48
botocore==1.17.48
Django==3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
isort==5.4.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
